// src/redux/actions/types/loginActionTypes.ts
export enum ELoginActionTypes {
  LOADING = 'LOGIN/USER_LOADING_BEGIN',
  AUTHORIZED = 'LOGIN/AUTHORIZED',
  ERROR = 'LOGIN/ERROR',
  SET_USER_DATA = 'LOGIN/SET_USER_DATA',
  SET_SERVER = 'LOGIN/SET_SERVER_NAME',
  TOGLE_INFO_POPUP = 'LOGIN/TOGLE_INFO_POPUP',
  SET_SERVERS_LIST = 'LOGIN/SET_SERVERS_LIST',
  SET_TZ_LIST = 'LOGIN/SET_TZ_LIST',
}